requirements.txt
python-telegram-bot==22.7
fastapi
httpx
cachetools
python-dotenv
pytest
